moved connectors folder
